Loaded active profile and other profiles into the interface. Started developing the ability to save the configuration file

+++ src/Services/HostsManager.Services/HostServices.cs
@@ -6,27 +6,28 @@ using HostsManager.Application.Entities;
 using HostsManager.Application.Helpers;
 using HostsManager.Application.Interfaces;
 
-namespace HostsManager.Application.Services
+namespace HostsManager.Services
 {
-    internal class HostServices:IHostService
+    internal class HostServices : IHostService
     {
         private readonly DirectoryInfo _hostsDirectory;
         private readonly List<Hosts> _defaultHosts;
         private const string HostsFileName = "HOSTS";
         private const string OriginalHostsFileName = "HOSTS_ORIGINAL";
+        private const string ActiveProfileIdentifier = "##ACTIVE_PROFILE:";
 
         public HostServices(DirectoryInfo hostsDirectory)
         {
             _hostsDirectory = hostsDirectory;
-            _defaultHosts= ReadDefaultHosts();
+            _defaultHosts = ReadDefaultHosts();
         }
 
         private List<Hosts> ReadDefaultHosts()
         {
             var filePath = File.Exists(OriginalHostsFilePath) ? OriginalHostsFilePath : HostsFilePath;
             var fileTextList = File.ReadAllLines(filePath).ToList();
-            
-            var uncommentedTextList = fileTextList.Where(p => !p.StartsWith("#"));
+
+            var uncommentedTextList = fileTextList.Where(p => !p.StartsWith("#")).ToList();
 
             var defaultHosts = new List<Hosts>();
             foreach (var uncommentedText in uncommentedTextList)
@@ -42,30 +43,42 @@ namespace HostsManager.Application.Services
 
             return defaultHosts;
         }
-        public void CreateNewHostsFile()
+
+        internal string GetActiveProfileName()
         {
-            File.Move(HostsFilePath, OriginalHostsFilePath);
+            var headerLine = File.ReadLines(HostsFilePath).First();
+            if (headerLine.Contains(ActiveProfileIdentifier))
+                return headerLine.Split(ActiveProfileIdentifier)[1];
+            return null;
+
+        }
+
+        private void CreateNewHostsFile()
+        {
+            if (!File.Exists(OriginalHostsFilePath))
+                File.Move(HostsFilePath, OriginalHostsFilePath);
+            File.Create(HostsFilePath).Close();
         }
 
         public void SetProfile(string profileName, Profile profile)
         {
             CreateNewHostsFile();
-            var fileText = new List<string> {$"##ACTIVE_PROFILE => {profileName}"};
+            var fileText = new List<string> { $"{ActiveProfileIdentifier}{profileName}"};
             fileText.AddRange(profile.Hosts.Select(host => $"{host.Ip}    {host.Host}"));
-            fileText.Add($"#DEFAULT_HOSTS");
-            fileText.AddRange(_defaultHosts.Select(host => $"##{host.Ip}    {host.Host}"));
-            File.WriteAllLines(HostsFilePath,fileText);
+            fileText.Add($"###DEFAULT_HOSTS");
+            fileText.AddRange(_defaultHosts.Select(host => $"###{host.Ip}    {host.Host}"));
+            File.WriteAllLines(HostsFilePath, fileText);
         }
 
         public IReadOnlyList<Hosts> DefaultHosts() => _defaultHosts;
-        
+
 
         public void RollbackHostsFile()
         {
-            if(!File.Exists(HostsFilePath) || !File.Exists(OriginalHostsFilePath))
-                throw  new Exception("It's not possible to rollback in the actual state");
+            if (!File.Exists(HostsFilePath) || !File.Exists(OriginalHostsFilePath))
+                throw new Exception("It's not possible to rollback in the actual state");
             File.Delete(HostsFilePath);
-            File.Move(OriginalHostsFilePath,HostsFilePath);
+            File.Move(OriginalHostsFilePath, HostsFilePath);
         }
 
         private string HostsFilePath => $"{_hostsDirectory.FullName}\\{HostsFileName}";
